change code to fit the database
I've tried to improve the code and make it more consistent. However I've encountered an issue and didn't realise it was on the database side. While trying to fix it I've tried deleting the national id column and adding back but order of the columns can't be changed easily so I've reverted best as I can and updated the code to work with it